+++ CHANGELOG.md
@@ -15,6 +15,42 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Change placeholder, add title tag and check organization search input search
 
+## [1.35.1] - 2024-10-01
+
+## [1.35.0] - 2024-09-10
+
+### Added
+- Added `getAccount` query on OrganizationDetailsSellers component to get the account data
+
+## [1.34.3] - 2024-09-05
+
+### Fixed
+- Review and reorder translations for better maintainability
+
+## [1.34.2] - 2024-09-03
+
+### Fixed
+
+- Fix English and Spanish string for payment terms
+
+## [1.34.1] - 2024-08-21
+
+### Fixed
+
+- Update pagination selected option in Organization Details pages
+
+## [1.34.0] - 2024-08-20
+
+### Fixed
+- Implement pagination for sellers in Organization Details page
+
+## [1.33.1] - 2024-08-06
+
+### Fixed
+
+- Adds optional changing on organization request
+
+
 ## [1.33.0] - 2024-07-18
 
 ### Feat